requirements.txt
absl-py==0.7.0
aniso8601==6.0.0
astor==0.7.1
astroid==2.2.5
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
pip3 install "./dependencies/imageai-2.0.2-py3-none-any.whl"
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
nltk==3.4
numpy==1.16.1
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
protobuf==3.6.1
pylint==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tflearn==0.3.2
Theano==1.0.4
typed-ast==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
